🏷️ | type(027): AllComponentsList interface and export in new AllComponents module
