+++ packages/client/src/game/ecs/components/all/AllComponents.ts
@@ -1,0 +1,17 @@
+import {
+  InputComponent,
+  KeymapComponent,
+  MatchConfig,
+  MovementComponent,
+  VelocityComponent,
+} from '../';
+
+interface AllComponentsList {
+  matchConfig?: MatchConfig;
+  keymap?: KeymapComponent;
+  input?: InputComponent;
+  movement?: MovementComponent;
+  velocity?: VelocityComponent;
+}
+
+export type { AllComponentsList };
