+++ packages/client/src/game/ecs/components/all/index.ts
@@ -1,0 +1,1 @@
+export * from './AllComponents';
