+++ packages/client/src/game/ecs/components/index.ts
@@ -3,3 +3,4 @@ export * from './keymap';
 export * from './input';
 export * from './movement';
 export * from './velocity';
+export * from './all';
